Improved path checking
If the file/directory exists by that path, then the path is valid

+++ src/main/java/me/theentropyshard/crlauncher/utils/FileUtils.java
@@ -56,6 +56,11 @@ public final class FileUtils {
     public static boolean isPathInvalid(String path) {
         try {
             Path file = Paths.get(path).toAbsolutePath();
+
+            if (Files.exists(file)) {
+                return false;
+            }
+
             FileUtils.createFileIfNotExists(file);
             FileUtils.delete(file);
         } catch (InvalidPathException | IOException e) {
